add support for execution hooks
New Features:
plan and destroy support both pre, and post execution hooks. There are
many times where terraforms `execute` statement are not sufficient, and
it will -always- trigger a change which is undesirable.

This new feature works by looking in the `hooks` directory in a
definition. If the hooks directory exists, and contains scripts that are
executable it will run them. The scripts -must- be named in accordance
to the pattern below:

./hooks/pre_apply.*
./hooks/post_apply.*
./hooks/pre_destroy.*
./hooks/post_destroy.*

Any file extension is supported, as is no file extension.

This feature also provides an added advantage to the scripts in that it
puts all terraform variables defined in the worker configuration, and
all remote variables within reach of the script. The following
environment variables are set:

TF_VAR_<variable_name>=<variable_value>
TF_REMOTE_<definition_name>_<variable_name>=<remote state value>

In addition two arguments are passed to the scripts, phase (pre/post)
  and command (apply/destroy). This would allow a single script with
  proper symlinks from the other names to handle all of the logic
  instead of needing to create specific scripts for different
  phases/operations.

Extra CLI Options for `terraform` command:
--force-apply: run terraform apply even if plan doesn't change, will
allow hooks to be force triggered if they are desired even without
change
--b64-encode-hook-values: this causes the worker to base64 encode all
terraform variables and state values that it sets in the hook execution
environment, this could be useful for handling complex terraform data
type outputs