upstream forked changes (#35)
* Add docker authentication

* add support for custom plugins

In order to support the use of custom plugins an alternative to the
hashicorp registry path needs to be provided. This PR adds support for
both custom URLs and filenames. It also adds validation that a version
is included in the plugin, which is required but was not previously
validated.

Adds tests for the following conditions:
- plugin is specified in the standard fashion and a version is provided
- plugin is specified but does not have a version
- plugin is specified but supplies a non-standard base URL
- plugin is specified but supplies a non-standard filename
- plugin is specified but supplies both a non-standard name and base URL

* removed debug prints

* update config fixture with jinja values

* update plugin handling

* fix plugin handling, add tf13/12 tests

* fix formatting

* fix formatting

Co-authored-by: Eric Bellefontaine <eric.bellefontaine@objectrocket.com>
Co-authored-by: David Grier <seraphin@gmail.com>