Increase the limit of deploymment names to 32 characters.
The limit on deployment name length was largely an arbitrary decision and was used since it was expected to use the {{ deployment }} variable it creates for inclusion in naming terraform resources. This increases the arbitrary limit but does not completely remove limits because unbounded deployment lengths will cause problems when setting up remote backends.