refactor the s3 backend
This refactoring included breaking down several of the long methods in
the s3 backend to smaller functions in order to enhance unit testing and
ensure that this section of code remains solid.

This also updates moto from v4.x to v5.x to take advantage of new
functionality and simplified use.

There are many test additions for the s3 backend, but also a lot of
changes to try and reduce the reliance on complicated fixtures and
provide smaller more consise tests.