update deps to address security issues
This not only updates security specific deps, but does a broad update of
all deps. The only problem areas:

- awscli: This requires an old colorama, but awscli isn't actually a
dep, so remove it. It was only included to make running the AWSCLI
easier at a point earlier in the projects life

- pytest-lazyfixtures: This pytest plugin relies on deprecated
functionality and has not been updated in >2 years, so likely will not
be fixed soon. The only tests that relied on lazy fixutres were able to
be pretty quickly fixed by using the pytest request.getfixturevalue
method