Move the root options to a Pydantic Model
The long term is goal is to move all CLI configurations to models.
Currently attributes are slapped onto one big inherited object and make
it difficult to determine where something is coming from, and there is
significant coupling.

Moving the options into a base model allows for all validation and
conditional logic to be handled within the model, and then the model
can be passed to places where the configuration may be needed. This
will lessen the overuse of passing around big sets of kwargs and
taking out what is needed, and will ensure that the CLI options
are seamlessly available in all the different areas they are needed.

This starts with just the RootCommand, which takes the "global" or
top level options.