get providers mirror working after refactoring
there were multiple problems with the providers collection, fixed
also added always_apply and always_include options to the definition
model, the intent is to allow some definitions to always be included
even when a --limit is passed.

this also refines some of the terrorm helpers to use the log module
instead of spitting stuff out to stdout.