INF-5796: Make remote state retrieval more resilient
In the event where a hook script execution is required in order to make terraform
execute in order to show the state (such as when a hook script defines the value of
specific variables) there is a circular condition where the remote outputs can not
be determined by terraform directly, and still made available to the hook.

In order to work around this issue, when the two initial mechanisms to get remote
state items fail, then the state is acquired from the backend (currently only S3
supported), and the state of the remote reference is used directly to obtain the
values.