INF-6170: Add option to disable remote state var lookup by hooks
Update the definition model, adding a field: hooks_disable_remotes. This name was chosen to "group" hook options together, this is currently the only one, but perhaps in the future there will be more need to control hbehavior of how hooks are executed. A unit test has been added to ensure the proper behavior when setting this option on the definition.

This also does some general maintenance:
- update deps, some with critical security alerts
- fix linting errors (make lint)
- update formatting  (make format)