Default to not scanning the definition
A snyk scan of the definition is actualy quite problematic. First,
when terraform init runs, all referenced modules are downloaded,
this means snyk will recursively scan all modules. This sounds
great in theory, but in practice it means alarming on things that
might only potentially be misconfigured, but are not actually, it
means the .snyk ignores from each module is not included in the
ignores because only one .snyk file can be used, and it doesn't
make sense to try and aggregate all the different modules snyk files
that is fragile at best, additionally, a single module may be
downloaded many times by a single definition, that means the noise
will be amplified and unable to be controlled.

An actual scan of the plan output, driven by the .snyk file in the
definitions repo is the only one that makes sense, as the definition
is where we intend to gather business rules, such as IAM policies.

Furthermore, each tf-module repository is already gated with its own
snyk code scan, and each of those modules need to address snyk failures.
Any .snyk ignores in modules will not be honoroed by a plan scan, which
is good, a module should not control final decisions about
security validation.